# openapi.yaml
openapi: 3.1.0
info:
  title: Preloop AI API
  description: REST API for Preloop AI issue tracking management
  version: 0.1.0
paths:
...
  /api/v1/issues/{issue_id}/comments:
    get:
      tags:
      - Comments
      summary: List Issue Comments
      description: Get a list of comments for a specific issue. Requires authentication.
      operationId: list_issue_comments_api_v1_issues__issue_id__comments_get
      security:
      - bearerAuth: []
      parameters:
      - name: issue_id
        in: path
        required: true
        schema:
          type: string
          title: Issue Id
      - name: organization
        in: query
        required: true
        schema:
          type: string
          description: Organization identifier
          title: Organization
        description: Organization identifier
      - name: project
        in: query
        required: true
        schema:
          type: string
          description: Project identifier
          title: Project
        description: Project identifier
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Maximum number of comments to return
          default: 20
          title: Limit
        description: Maximum number of comments to return
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of comments to skip
          default: 0
          title: Offset
        description: Number of comments to skip
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Comments
      summary: Add Issue Comment
      description: Add a new comment to a specific issue. Requires authentication.
      operationId: add_issue_comment_api_v1_issues__issue_id__comments_post
      security:
      - bearerAuth: []
      parameters:
      - name: issue_id
        in: path
        required: true
        schema:
          type: string
          title: Issue Id
      - name: organization
        in: query
        required: true
        schema:
          type: string
          description: Organization identifier
          title: Organization
        description: Organization identifier
      - name: project
        in: query
        required: true
        schema:
          type: string
          description: Project identifier
          title: Project
        description: Project identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/comments/search:
    get:
      tags:
      - Comments
      summary: Search Comments
      description: 'Search for comments using full-text or similarity search.

        Requires authentication and checks user access to related issues/projects.'
      operationId: search_comments_api_v1_comments_search_get
      security:
      - bearerAuth: []
      parameters:
      - name: query
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search query text for comment body or vector search
          title: Query
        description: Search query text for comment body or vector search
      - name: search_type
        in: query
        required: false
        schema:
          type: string
          description: Type of search to perform ('fulltext' or 'similarity')
          enum:
          - fulltext
          - similarity
          default: fulltext
          title: Search Type
        description: Type of search to perform ('fulltext' or 'similarity')
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Maximum number of comments to return
          default: 10
          title: Limit
        description: Maximum number of comments to return
      - name: issue_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter comments by a specific issue ID (UUID)
          title: Issue Id
        description: Filter comments by a specific issue ID (UUID)
      - name: project_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter comments by parent issue's project ID (UUID)
          title: Project Id
        description: Filter comments by parent issue's project ID (UUID)
      - name: organization_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter comments by parent issue's organization ID (UUID)
          title: Organization Id
        description: Filter comments by parent issue's organization ID (UUID)
      - name: author
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter comments by author (username)
          title: Author
        description: Filter comments by author (username)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentSearchResults'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CommentCreate:
      properties:
        body:
          type: string
          title: Body
          description: Comment text
        meta_data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta Data
          description: Additional comment metadata
      type: object
      required:
      - body
      title: CommentCreate
      description: Model for creating a new comment.
    CommentList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CommentResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: CommentList
      description: Response model for a list of comments with pagination details.
    CommentResponse:
      properties:
        body:
          type: string
          title: Body
          description: Comment text
        meta_data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta Data
          description: Additional comment metadata
        id:
          type: string
          title: Id
          description: Comment unique identifier
        issue_id:
          type: string
          title: Issue Id
          description: Issue ID this comment belongs to
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
          description: Comment author
        created_at:
          type: string
          title: Created At
          description: Creation timestamp
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
          description: Last update timestamp
        score:
          anyOf:
          - type: number
          - type: 'null'
          title: Score
          description: Similarity score for search results (if applicable)
      type: object
      required:
      - body
      - id
      - issue_id
      - created_at
      title: CommentResponse
      description: Response model for comment data.
    CommentSearchResults:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CommentResponse'
          type: array
          title: Items
          description: Search result items
        total:
          type: integer
          title: Total
          description: Total number of matching comments
        query:
          type: string
          title: Query
          description: Search query used
      type: object
      required:
      - items
      - total
      - query
      title: CommentSearchResults
      description: Response model for comment search results.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication